pyproject.toml
[project]
name = "fletxbuild"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fletx>=0.2.0",
    "sqlalchemy",
    "flet-audio==0.1.0",
]